refactoring

+++ src/main/java/org/erachain/gui/models/UnconfirmedTransactionsTableModel.java
@@ -144,7 +144,7 @@ public class UnconfirmedTransactionsTableModel extends TableModelCls<Long, Trans
     }
 
     @Override
-    public int getMapSize() {
+    public long getMapSize() {
         return map.size();
     }
